Split the test into two classes.
