+++ src/NServiceBus.AcceptanceTests/NServiceBus.AcceptanceTests.csproj
@@ -73,10 +73,11 @@
     <Compile Include="Routing\When_configure_routes_for_unobtrusive_messages.cs" />
     <Compile Include="Routing\When_extending_command_routing.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\When_extending_event_routing.cs" />
+    <Compile Include="Routing\When_registering_publishers_unobtrusive_messages_config.cs" />
     <Compile Include="Routing\When_sending_non_message_with_routing_configured_by_type.cs" />
     <Compile Include="Routing\When_sending_non_message_with_routing_configured_by_assembly.cs" />
     <Compile Include="Routing\When_sending_non_message_with_routing_configured_via_mappings.cs" />
-    <Compile Include="Routing\When_registering_publishers_unobtrusive_messages.cs" />
+    <Compile Include="Routing\When_registering_publishers_unobtrusive_messages_code.cs" />
     <Compile Include="Routing\When_publishing_an_interface_with_unobtrusive.cs" />
     <Compile Include="Routing\When_replying_to_message_with_interface_and_unobtrusive.cs" />
     <Compile Include="Routing\MessageDrivenSubscriptions\When_subscribing_to_a_base_event.cs" />
